backend/requirements.txt
anyio==3.6.2
autopep8==2.0.0
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.87.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pycodestyle==2.9.1
pydantic==1.10.2
pymongo==4.3.3
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
